Re-export remaining interfaces
